Generalise Overview component
